+++ CHANGELOG.md
@@ -84,7 +84,7 @@ Release 0.3.0
 =============
 
 - update dependencies
-- added node-based YAML merge function `MergeYAMLNodes` to prevent 
+- added node-based YAML merge function `MergeYAMLNodes` to prevent
   deserialization inaccuracies
 - added `FromFiles` to facilitate reading from multiple input files
   that overlay each other in order